[usb transport] Straighten out BulkReaderThread locking
The thread was using a QMutex for control flow, which
doesn't work in Qt5 because in Qt5 you're not allowed to
lock a mutex in one thread and then unlock it in another.

Switched to using a QWaitCondition.

Also, for efficiency, switched to allocating just one
buffer and then trading it back and forth, instead of
allocating and freeing a new one for every packet.