[mtp] Abort segmented getObject response if session was closed
If a session was closed (for example by unplugging the USB) while
sendObjectSegmented() was active, it would keep trying to send more
segments every time the connection was re-established. This interfered
with attempts to open a new session.

Fixing this will make it less likely that MTP interruptions will require
a restart to fix. Now the segment-sending loop will exit if it detects
that the session has closed or the current transaction has been canceled.