[storageplugin] fix crash in DeleteObject with handle=ALL
The deleteItem handler tried to loop over a QHash while
removing items from it, and this doesn't work unless you use
special calls for the removing. (Which is awkard since the
removal is done in deleteItemHelper).

Fixed by looping over a copy.

This bug was exposed by the storage-test unit test.