Change StorageID assigning as per Richard Braakman's suggestion
This should perhaps be assignStorageId(1, i + 1). In the commented-out
code that handles multiple plugins, the first argument is the plugin
number. The second one is called partitionNo. I'm not sure it matters
anywhere, but I guess that conceptually the home dir and the sd card are
partitions of the fsstorageplugin.