Let Tracker extract name of the song from file-embedded tags
Some initiators send media file properties with Name equal to Object
File Name, even though the tags embedded in the file may contain a value
more suitable for display in a player as a song or movie title than
a mere filename.

Thus, when we encounter a filename in the Name property, we don't store
the value and let Tracker try extracting the name from the file's
embedded metadata.