[storage] move StoragePlugin::m_storagePath to FsStoragePlugin
This protected class member doesn't play any role in the abstract
interface of a storage plugin. In general, for some potential kinds of
storage plugin, implemented in the future, the concept of "storage path"
might not even make any sense. Consequently, let's move the member to
the place of its use, which is FsStoragePlugin.