[fsstorageplugin] pass ObjHandle by value
Changed methods removeFromStorage and deleteItemHelper to take
their ObjHandle by value instead of by reference, on the grounds
that they're going to remove that handle from some maps and this
is highly awkward if the calling function passed in a reference
taken from those maps.

Specifically, the autotests found an error where handleFSMove
passes m_pathNamesMap[oldPath] as the argument, and the []
operator returns a reference to the map element, which is
then deleted in removeFromStorage.

This is the kind of subtle bug that makes C++ so beloved.