[deviceinfo] Use SSU modelname to override MTP friendlyname. Fixes JB#33462
We want to make the friendly name exposed over mtp to be device specific.
The name is defined in xml configuration file where the rest of the data
is not device specific.

Instead of maintaining a number of almost identical configuration files
in adaptation packages, we can query the appropriate name from SSU and
override the friendly name specified in the xml configuration.