[threadio] Cap USB bulk write lengths. Fixes JB#33480
Depending on kernel side implementation, usb bulk writes that are too
large to fit into buffers at once can be accepted partially or rejected
altogether. In the latter case sending of larger mtp packets will fail
systematically - one of the 1st symptoms of that happening is that the
device icon does not show up on the pc end because the icon data is
large enough to cause problems.

Limit length of individual writes to 64KB. In case of EIO errors, halve
the limit until writes succeed.

So far the observed write lengths that are accepted have been 16KB long,
so this should not have negative effect on transfer speeds.