[thumbnailer] Do not block at Queue D-Bus call. Fixes JB#34779
While the actual thumbnailing is handle fully asynchronously, buteo-mtp
blocks and waits for org.freedesktop.thumbnails.Thumbnailer1.Queue method
call replies. While individual D-Bus round trips are relatively quick, it
occurs on critical path and can cause significant delays to reaching the
storageReady condition - especially if there are lots of images that
are missing thumbnails.

Use pending call watcher to handle the reply message without blocking.

Also revise the bookkeeping so that only images that are successfully
passed to thumbnailer are inserted into the uri map.

Remove artificial 2000 thumbnails limit as it is not justified in any
way and does not probably work as intended: Thumbnail requests are still
made - to be ignored by buteo-mtp when finished.