[thumbnailer] Request thumbnails in batches. JB#34779
If there are multitude of images that are missing a thumbnail, sending
each file individually to thumbnailer creates a lot of dbus related io
for the mainloop to process. Additionally each finished thumbnail causes
object property changes which then needs to be sent over to the host,
which can cause problems during starup when the host is probing the
device.

Use timer to delay thumbnail requests so that they can be issued in
batches.

Use longer delay on startup to make it more likely that the device
reaches storageReady state and host manages to open mtp session before
thumbnail related property changes start to trickle in.