[mtptransporterusb] Wait for storageReady before accpting mtp ops. JB#34779
Use of freely running bulk reader thread and non-synchronized responder
state changes during bulk writes combined with attempts to delay execution
of mtp requests received before storageReady conditions is reached create
a timing hazard that can make it look like there are two overlapping bulk
writes - the latter one is skipped, which leaves the host waiting for a
reply that never comes.

Assume that the delayed execution was needed because reaching storageReady
used to take so long and keep bulk reader thread stopped until the device
is ready to serve all mtp requests, not just the ones that do not need
the storage plugin.