[thumbnailer] Remove stale thumbnails to avoid ipc ringing. JB#35652
If a situation arises where buteo-mtp thinks a thumbnail is not up to date
while thumbnailer thinks it is, this results in buteo-mtp perpentually
requesting a new thumbnail to be created. Possible reasons for this
happening might include things like having the initial thumbnailing request
be made while transfer of an image file has not yet been completed.

Attempt to remove old thumbnail file before requesting a fresh one to be
created.