[fsstorageplugin] Do not send events about unprobed objects. Fixes JB#35652
Generating thumbnails causes object/property changed events to be sent. If
the mtp has not been used for a while and has large number of pictures
without thumbnail data, the host gets avalanche of events while attempting
to do initial storage probing - which can cause problems in form of delays
and timeouts.

Also in general it makes no sense to send change data about objects that are
not yet known to the host side.

Keep track of objects that the host as queried about and send object /
property changed events only for those objects.