[fsstorageplugin] Do not bypass image checks in property queries. JB#35652
Unlike the rest of the thumbnail related logic, the property value caching
does not use isImage() predicate to check whether an object can have a
thumbnail or not. As a consequene explicit thumbnail requests made by
client result in issuing thumbnailing requests that will systematically
fail.

Also the name of the predicate is a bit misleading because it is used for
detecting images that the thumbnailer can be assumed to handle successfully
rather that telling apart image files from non-image files.

Change the predicate name to isThumbnailableImage() and use it also in
property value caching.

Improve diagnostic logging of reasons for not providing thumbnail data.