[fsstorage] Avoid creating files with privileged gid. Fixes JB#36452
The mtp functionality is implemented as msyncd plugin. The msyncd
process is running with "privileged" effective group id. This leads
to all files uploaded to the device via mtp getting assigned to
"privileged" group - which is not as intended.

When creating new files, explicitly set the file ownership so that
the real gid (= "nemo") gets used instead of "privileged".

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>