[mtprxcontainer] Fix string deserialization. JB#43270
Character data is used as-is from line buffer and result value is not
updated when null/empty string is received.

Utilize deserialize() method for character data so that it works as
expected also when host endianess differes from line protocol.

If empty/null string is found, set result value to empty string.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>