[descriptor] Use compilation time constant byteswaps. Fixes JB#44871
After gcc upgrade use of htoleNN() macros leads to "initializer element
is not constant" errors.

As all the values we need to turn into little endian format are constants,
we can utilize __bswap_constant_NN() to keep the swapped values compilation
time constant.

As this leads to less optimal code when dealing with non-constant values,
move the macros from header to source in order to avoid accidental use
in other modules in the future.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>