[fsstorageplugin] Do not trash files in recursive scanning. Fixes JB#44884
If directory with files is seen by buteo-mtp to show up as inotify event
(due to atomic move of already populated directory, or making a symlink
to such directory), all the files within such directories are written
to and truncated due to reversed arguments in addToStorage() recursion.

Correct the order of sendEvent and createIfNotExist parameters.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>