[tracker] Disable metadata writes from MTP to tracker. Fixes JB#51156
When files are transferred onto the device via MTP, clients can optionally
supply metadata for those files. This data is written directly to
Tracker's database, creating entries that would normally be created
by tracker-miner-fs and tracker-extract. There is code intended to tell
Tracker not to index the new files, but it has never been enabled in
Sailfish, it if it ever was.

With the new Tracker upgrade to 2.3.4, this behaviour has caused problems:
 - There are several errors and inconsistencies with tracker in the
 supplied data and how it is stored. This code has been unmaintained for
 several years.
 - Tracker tries to write its own data, but doesn't recognize the
 MTP sourced data because it intentionally ignores anything it's not
 the source for. This causes uniqueness errors in the backend which breaks
 indexing and further updates.

This commit disables the storage of metadata by MTP, but does not remove
the code. Tracker data is still read and made available to clients. This
leaves a lot of dead code that should be cleaned up at some point once
we're sure this is the way to go.

Signed-off-by: Andrew Branson <andrew.branson@jolla.com>