[fsstorageplugin] Get disk labels from UDisks. Fixes JB#53587
Memory cards mounting was recently changed so that mountpoint is
  /run/media/<USER>/<UUID>
instead of
  /run/media/<USER>/<LABEL>
and this causes memory cards to show up over mtp with long hex-only
names that are not really meant for human consumption.

Query block device information from UDisks2 D-Bus service, and
use disk label from block device associated with the mount point
when available.

For those mount points for which label information is not available,
enumerated "Card N" labels are used.

Synchronous D-Bus IPC is used as this is needed in procedural logic
executed before buteo-mtp enters mainloop.

As buteo-mtp does not use D-Bus SystemBus for other purpose, and
attempt is made to close connection after it is not needed anymore.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>