Remove tracker metadata copy/move functionality
This seems mostly unnecessary, new files or moved paths should anyway
trigger reindexing with more robust end result than trying to do it all
manually. And depending on the tracker configuration could be that
indexing is not even used on the new path, in which case the content
shouldn't end up in the tracker databases. And even further, we have
been doing some tracker resets on updates which has meant that only
the mined data has had a chance of surviving.