[tests] Fix deviceinfoprovider_test failures. JB#52789
Vendor extension check fails due to whitespace differences between xml
configuration file content and what buteo-mtp code returns. Based on
MTP specification V1.0
  The name of the extension set is terminated with the colon character,
  and the version number is terminated with the semicolon character. A
  single space is required after each terminator character (':' and ';').
the xml data is wrong and needs to be fixed.

Supported formats checks fail because xml configuration data has not been
used as-is in ages. Change test code so that it verifies that formats
listed in xml are supported but does not assume that actually supported
formats are 1:1 with xml configuration.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>